Refactor Job and its compiler.
