+++ compiler/src/main/scala/com/asakusafw/spark/compiler/SparkClientCompiler.scala
@@ -47,7 +47,8 @@ import com.asakusafw.lang.compiler.planning.Plan
 import com.asakusafw.spark.compiler.graph.{
   BranchKeysClassBuilder,
   BroadcastIdsClassBuilder,
-  Instantiator
+  Instantiator,
+  JobCompiler
 }
 import com.asakusafw.spark.compiler.planning.SparkPlanning
 import com.asakusafw.spark.compiler.spi.{
@@ -141,19 +142,18 @@ object SparkClientCompiler {
     def branchKeys: BranchKeysClassBuilder
     def broadcastIds: BroadcastIdsClassBuilder
 
-    def options: CompilerOptions
-
-    def nodeCompilerContext: NodeCompiler.Context
-    def instantiatorCompilerContext: Instantiator.Context
+    def jobCompilerContext: JobCompiler.Context
   }
 
   class DefaultContext(val flowId: String)(jpContext: JPContext)
     extends Context
+    with JobCompiler.Context
     with NodeCompiler.Context
     with Instantiator.Context
     with OperatorCompiler.Context
     with AggregationCompiler.Context {
 
+    override def jobCompilerContext: JobCompiler.Context = this
     override def nodeCompilerContext: NodeCompiler.Context = this
     override def instantiatorCompilerContext: Instantiator.Context = this
     override def operatorCompilerContext: OperatorCompiler.Context = this
