+++ compiler/src/test/scala/com/asakusafw/spark/compiler/MockCompilerContext.scala
@@ -29,7 +29,9 @@ import com.asakusafw.lang.compiler.model.description.ClassDescription
 import com.asakusafw.lang.compiler.model.info.{ ExternalInputInfo, ExternalOutputInfo }
 import com.asakusafw.spark.compiler.graph.{
   BranchKeysClassBuilder,
-  BroadcastIdsClassBuilder
+  BroadcastIdsClassBuilder,
+  Instantiator,
+  JobCompiler
 }
 import com.asakusafw.spark.compiler.spi.{
   AggregationCompiler,
@@ -54,6 +56,59 @@ class MockCompilerContext(val flowId: String)
 }
 
 object MockCompilerContext {
+
+  class JobCompiler(val flowId: String)(jpContext: JPContext)
+    extends JobCompiler.Context
+    with NodeCompiler.Context
+    with Instantiator.Context
+    with OperatorCompiler.Context
+    with AggregationCompiler.Context {
+
+    override def nodeCompilerContext: NodeCompiler.Context = this
+    override def instantiatorCompilerContext: Instantiator.Context = this
+
+    override def operatorCompilerContext: OperatorCompiler.Context = this
+    override def aggregationCompilerContext: AggregationCompiler.Context = this
+
+    override def classLoader: ClassLoader = jpContext.getClassLoader
+    override def dataModelLoader: DataModelLoader = jpContext.getDataModelLoader
+    override def options: CompilerOptions = jpContext.getOptions
+
+    override val branchKeys: BranchKeysClassBuilder = new BranchKeysClassBuilder(flowId)
+    override val broadcastIds: BroadcastIdsClassBuilder = new BroadcastIdsClassBuilder(flowId)
+
+    override def getInputFormatInfo(
+      name: String, info: ExternalInputInfo): Option[InputFormatInfo] = {
+      Option(InputFormatInfoExtension.resolve(jpContext, name, info))
+    }
+
+    private val externalInputs: mutable.Map[String, ExternalInputReference] = mutable.Map.empty
+
+    override def addExternalInput(
+      name: String, info: ExternalInputInfo): ExternalInputReference = {
+      externalInputs.getOrElseUpdate(
+        name,
+        jpContext.addExternalInput(name, info))
+    }
+
+    private val externalOutputs: mutable.Map[String, Unit] = mutable.Map.empty
+
+    override def addExternalOutput(
+      name: String, info: ExternalOutputInfo, paths: Seq[String]): Unit = {
+      externalOutputs.getOrElseUpdate(
+        name,
+        jpContext.addExternalOutput(name, info, paths))
+    }
+
+    override def addClass(builder: ClassBuilder): Type = {
+      for {
+        os <- managed(jpContext.addClassFile(new ClassDescription(builder.thisType.getClassName)))
+      } {
+        os.write(builder.build())
+      }
+      builder.thisType
+    }
+  }
 
   class NodeCompiler(val flowId: String)(jpContext: JPContext)
     extends NodeCompiler.Context
